feat(utils): refactored utils to internal directory, following standard go project format
